requirements.txt
cassandra-driver
langchain
openai
pypdf
cassio>=0.1.1
tiktoken==0.4.0
streamlit